Delete security and fix room price null

+++ src/main/java/usth/edu/accommodationbooking/model/Room.java
@@ -20,7 +20,7 @@ public class Room {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String roomType;
-    private BigDecimal roomPrice;
+    private Integer roomPrice;
     private boolean isBooked = false;
     @Lob
     private Blob photo;
