+++ src/main/java/usth/edu/accommodationbooking/service/IRoomService.java
@@ -4,10 +4,9 @@ import org.springframework.web.multipart.MultipartFile;
 import usth.edu.accommodationbooking.model.Room;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.sql.SQLException;
 
 public interface IRoomService {
 
-    Room addNewRoom(MultipartFile photo, String roomType, BigDecimal roomPrice) throws SQLException, IOException;
+    Room addNewRoom(MultipartFile photo, String roomType, Integer roomPrice) throws SQLException, IOException;
 }
